refactor(decompose): split webapp/src/shell/edge/main/cli/commands/node/view.test.ts into folder
